% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Linreg Object}
\usage{
\method{plot}{linreg}(linreg)
}
\arguments{
\item{linreg}{linreg object containing coefficients B and regularization parameter reg.}

\item{X}{A numeric matrix of predictor variables.  Each row represents an observation, and each column represents a predictor.}

\item{Y}{A numeric vector of response variables.}
}
\value{
ggplot plot.
}
\description{
Plots histogram of residuals.
}
